mpu: modify copy middleware to support uploadPartCopy
An uploadPartCopy operation will have part-number and upload-id
parameters. These must be removed from a copy source GET request in
the copy middleware which precedes mpu middleware. They are only
relevant to the copy target PUT request and would cause the mpu to
treat the copy source GET request as a listParts operation.

Change-Id: I1e560a2d39a647ef8dbf3f85e2563894f2f94fea