mpu: add x-backend-etag-is-at header to GET/HEAD requests
When an MPU is completed, its mpu-etag is written as a sysmeta item of
the manifest object. When a client GET or HEAD request is handled, the
mpu-etag is copied to the Etag header in the client response. When a
GET or HEAD request with an If-[None-]Match header is handled, the
object server must be directed to look for the correct etag to compare
with the header value by adding an X-Backend-Etag-Is-At header with the
name of the sysmeta item.

Change-Id: I7a4f988c277841a4d79e6aafb29605fdf9de98d0