mpu: detect content-type during createUpload when not given
MPU middleware must auto-detect the content-type of an mpu if it is not
given. This cannot be delegated to the proxy app because:

(a) When the symlink is PUT the symlink middleware will set the
content-type to application/symlink if it is not otherwise
specified. This content-type then appears in the container listing for
the mpu.

(b) When the manifest is PUT its name ends with the upload id so any
extension that the proxy app could use to guess the content-type is
obscured. The proxy will therefore always add thr default content-type
of application/octet-stream to the manifest PUT. This is the
content-type that will be returned for a GET/HEAD of the mpu.

Change-Id: I217855207e9d85f304c27f0fcaeeb1f33a8cad41