mpu: fix last-modified time in UploadPartCopy response
Previously, the s3api multi_upload PartController would add an
X-Timestamp header to an UploadPartCopy request, and then use the
rounded value of that timestamp for the Last-Modified time in the
response body. However, mpu middleware was not forwarding the s3api
request x-timestamp with the PUT subrequest to the parts container, so
the actual rounded modified time of the part might be different if the
request handling spanned a transition from one second to the next.

This was intermittently detected by some assertions in
test/functional/s3api/test_multi_upload.py which compared the
last-modified times of UploadPartCopy responses to the values returned
in a ListParts response.

Two changes are made, either of which addresses the problem:

1. the s3api multi_upload PartController no longer adds x-timestamp to
the UploadPartCopy request, and now uses the response Last-Modified
header to populate its response body. It shouldn't be necessary for
the s3api to set x-timestamp.

2. The mpu middleware will now forward an x-timestamp header from an
UploadPart request with the parts container subrequest. If an earlier
middleware has a reason to set the x-timestamp then it should be
forwarded. Note that mpu middleware itself will set x-timestamp on all
requests if it is not already present.

Change-Id: I6f7cd7514be25bb5996a352deadd2e029e1bf24d