mpu: ensure redirected POST headers are unencrypted
The encypter middleware transforms user object metadata headers to
encrypted transient sysmeta header when a POST is sent to an mpu
symlink. The POST is redirected to the manifest by mpu middleware,
which needs to ensure that any user metadata headers with the original
request are sent to the manifest. The encrypted transient metadata
created when the POST was sent to the symlink cannot be decrypted when
the manifest is read by a GET or HEAD because the keys used for
encryption and decryption are specific to the object path.

Change-Id: Icdbc2b6d885372b4831845b8222e00082ce5bf3f