mpu: store max part number in manifest in sysmeta
The maximum part number referenced in the manifest is stored as
sysmeta with the manifest.

The part numbers referenced in the manifest do not have to increment
by 1 i.e. there can be gaps. It may be useful in the future to know
the maximum part number that was referenced in the manifest. For
example, if this is the same as the number of parts, then the path to
a part can be inferred from the part number without having to read the
manifest.

Change-Id: I36428e03ebd874892095e13a81e722a3d5550c2c