mpu: repeated completeUploads get unique timestamp offsets
When an MPU is completed, the manifest is PUT using the timestamp at
which the MPU session was created. This satisfies the requirement that
an MPU is overwritten by a request later than the time at which the
session was created.

However, it is generally a bad idea to have two independent PUT
requests to the same path using the same timestamps, which could
happen if there are concurrent or retried completeUpload requests. The
timestamp used with manifest PUTs is therefore set to the created
timestamp with an offset equal to the delay to the client's
completeUpload request timestamp. That ensures that each
completeUpload request uses a unique timestamp for the manifest, but
they all sort earlier than a later 'normal' timestamp.

The object-versioning middleware is modified to ignore offsets in
timestamps used to construct version-ids.  The normal part of the
timestamp is sufficient to differentiate versions.

Change-Id: Ibab4544013f926dc4c3603010e8c40a158c2185e