mpu: use object event history to detect orphans
Maintain a container with history of events for every object. Teach
mpu auditor to detect and clean up deleted obsolete versions based on
the history.

TEMP:
* workaround to fix replicator probe tests re unexpected async pending

Maintain a history of objects events in a hidden 'history'
container. The history container is updated by additional updates
after object PUTs and DELETEs to the user namespace container.

The mpu-auditor periodically inspects the history container. Every
null version of an object older than the latest null version is
considered obsolete. Also, a retained version is made obsolete by a
delete event for the same version.

The name of each history version encapsulates the version time. If it
is an mpu, the history entry systags has the metadata required to
locate mpu parts.

Have the proxy add container host/device/partition info to custom
container updates so that the object server can send the updates in
sync with requests.

Change-Id: I2f9eeff93b2ec27a28d7cc6e16db335328949dd7
Signed-off-by: Alistair Coles <alistairncoles@gmail.com>