tests: rename zuul jobs that test s3api
The swift-tox-func-s3api-tests-tempauth runs the tests in test/s3api.
The swift-tox-func-s3api-ceph-s3tests-tempauth doesn't run the tests
in test/s3api. This patch changes the names of those jobs to
disambiguate the prefix 'swift-tox-func-s3api' and avoid any
suggestion that the second job is a super-set of the first job.

Change-Id: I323cf7e32200a2f895f0b051454b00047f7de300