tests: Fix test_LoggerFileObject_recursion
It used to be that when a logger encountered an error trying to handle
/ emit a message, it would attempt to log about the error via the same
logger. Long ago, this could lead to infinite recursion in Swift's
LoggerFileObject, so we fixed it and added a test that included
showing the bad stdlib behavior.

Recently, stdlib did a similar fix so the recursion doesn't happen. See

 - https://github.com/python/cpython/issues/91555 and
 - https://github.com/python/cpython/pull/131812

Of course, now our test breaks (at least, on 3.13.4). Relax the
assertion a little.

Related-Change: Ia9ecffc88ce43616977e141498e5ee404f2c29c4
Change-Id: Id2f490d4204b8eaf07857cb84ed783bec19b8511