Add replacement for eventlet.Timeout
Replace all direct imports of eventlet.Timeout with a new Timeout class
in swift.common.concurrency, re-exported and subclassed by
swift.common.exceptions. eventlet.Timeout is both used as an exception
as well as a context manager to timeout calls within the context block.

This is not available without eventlet; in this case timeouts are either
enforced by using timeout parameters in existing functions (like
http_connect) or using socket timeouts. This requires using the socket
as argument to the Timeout class; if eventlet is used it will be
filtered out.

Change-Id: Ic0bf3fadc51d75b436b05ded9f0fe8209c7e8ecf
Signed-off-by: Christian Schwede <cschwede@mailbox.org>