WIP: Add gunicorn as wsgi server
Simple replacement using gunicorn as a wsgi server instead of eventlets
builtin server. This commit only adds a basic gunicorn wsgi server which
can be used if the environment setting "DISABLE_EVENTLET=true" is set.

Eventlet dependencies are not removed with this commit, and in fact both
gunicorn and eventlet are running together (though not performant).

Change-Id: Ia47a23dfafe193527506771d351ada806eed69b7
Signed-off-by: Christian Schwede <cschwede@redhat.com>