Add replacement for eventlet.tpool
eventlet uses tpool.execute(func) to run the func in a threadpool,
allowing the current coro to yield so other coros can run in the
meantime. When running without eventlet, there is no need to run func in
a separate thread because it is already running within a thread.

There is a difference though: running in a threadpool limits the number
of running tpool.execute() to the size of the threadpool. This limit
does not exist when executing directly.

Change-Id: I33f5a9a698b1de5f4b778e28645a9af9d9af6cd2
Signed-off-by: Christian Schwede <cschwede@mailbox.org>