mpu: use relic and action rows to drive audit
Detect when mpu's are made obsolete and generate relic rows in the
object table. Also generate action rows in a new action table, which
the mpu auditor uses to find cleanup work to do.

Don't create relics/actions when an unversion is superseded by a
version. Instead, keep track of the latest unversion in a hidden row.

Change-Id: Id200cb35b2935cbeee50d6f93e084f76cc348188
Signed-off-by: Alistair Coles <alistairncoles@gmail.com>