Device in fresh db id suffix
Currently there is a bug, that when we shard a container, the fresh db
that is created doesn't append the device as a suffix to it's db id.
We use this to help us track whose syncing with who in incoming/outgoing
syncs.

The reason it wasn't being added is because we'd build the freshdb in
<dev_path>/tmp/fresh<uuid4>.db which would then be renamed over to the
fresh db name ready to use. Hoever, this path is much too short and the
get_device_path function expected the db to be keeper in the filesystem,
so would fail to find it.

This patch moved the tmp location to be in the db_dir and use the .tmp
notation used elsewhere in the database broker initialize method. This
keeps the tmp db file far enough into the file system for
get_device_path to work but also is using an existing naming and
location scheme.

Change-Id: I9887bc2bfdcc304430026b0e87076a3b62a2a2b8
Signed-off-by: Matthew Oliver <matt@oliver.net.au>