Recliam db_dir/*.tmp files
Currently the db_replicator will clean up files older then reclaim_age
in the device_tmp directory, but it doesn't clean up any old .tmp files
located in the db_dir directory. These .tmp files are created in
initialize broker method, which will also then do a rename to the final
db file name. We've found a bunch in production along side the db files
themselves so there must be an edgecase where it's not renamed, maybe a
race condition with replication or the PUT.

In anycase, we now want to move the sharding freshdbs to be initially
created in the same manner, <db_dir>/fresh<uuid4>.db.tmp and then rename
once all configured. So there is a chance that these might be left
around from time to time. So would be a good idea to clean them up.

This patch adds a new `_reclaim_tmp_dbs` method that is called from
`_reclaim` that will generate a list of *.tmp files which is then passed
to `unlink_paths_older_than`, cleaning up .tmp files that are older then
reclaim_age.

Change-Id: If30005269d40a1a3f711008b5d3b863efc9fb683
Signed-off-by: Matthew Oliver <matt@oliver.net.au>