TombstoneReclaimer: quote string timestamp in sql query
Reclaim age and sync "timestamp" args passed to
DatabaseBroker.reclaim() and TombstoneReclaimer.__init__() are
expected to be floats. This is adhered to in code, but many tests pass
a string representation of a Timestamp instance. The string has been
tolerated because it has so far always been identical to a
str(float).

However, if a Timestamp instance has an offset then its string
representation is not equivalent to str(float), and
TombstoneReclaimer.reclaim() would raise an error because the reclaim
age value in the SQL query is not quoted and does not appear to be any
recognised type.

This patch quotes the reclaim age in the TombstoneReclaimer SQL
query. This makes it tolerant of being passed string representations
of Timestamps with offsets. This is not yet expected, but,
nevertheless, text in SQL queries should be quoted.

Change-Id: Ic70689db1e4ddc6f526c283bf0cbb5862b16fd90
Signed-off-by: Alistair Coles <alistairncoles@gmail.com>
Co-Authored-By: Alistair Coles <alistairncoles@gmail.com>