sharder: use correct Timestamp formats
Use Timestamp.internal for X-Timestamp header when making a PUT
request to send shard range updates to the root container. The unit
test already asserts that the internal format is sent, but the test
failed to use a Timestamp where the internal and normal formats
differed..

Don't try to cast timestamp strings to floats when sorting
CleavingContext metadata. Timestamp strings may have the form
<float>_<offset> so it is safer to cast a Timestamp string to a
Timestamp and then to cast the Timestamp to a float when necessary.

Change-Id: I59e140d1226a4b0c2c8f37678bc2dc87d9ffa631
Signed-off-by: Alistair Coles <alistairncoles@gmail.com>
Co-Authored-By: Alistair Coles <alistairncoles@gmail.com>
Signed-off-by: Matthew Oliver <matt@oliver.net.au>