OldestAsyncPendingTracker accepts Timestamp.internal
The OldestAsyncPendingTracker.add_update() method is passed a
timestamp string which is derived from the name of the async_pending
file. This string may take the internal form of a Timestamp which
cannot be directly cast to a float; the string should be cast to a
Timestamp instance and then cast to a float.

Change-Id: Ibe48a9193cd0a062cbf629f2ff0d950881380f48
Signed-off-by: Alistair Coles <alistairncoles@gmail.com>